changed utils to seperate files, made commandPage to arrayPage and accepts all array types, fixed up some issues with registering global and guild commands
