+++ src/buttons/helpPage/helplastpage.ts
@@ -1,6 +1,7 @@
 import { ColorResolvable, Message, MessageActionRow, MessageButton, MessageEmbed } from "discord.js";
-import { capitalize, commandPaginate } from "../../utils/utils";
 import Buttons from "../../interfaces/buttons";
+import { arrayPage } from "../../utils/arrayPage";
+import { capitalize } from "../../utils/capitalize";
 import { deleteButton } from "../../globals";
 
 export const buttons: Buttons = {
@@ -17,7 +18,7 @@ export const buttons: Buttons = {
         let finalPage = 1;
         let notMax = false;
         while (!notMax) {
-            const cmds = commandPaginate([...client.commands.values()], 4, finalPage);
+            const cmds = arrayPage([...client.commands.values()], 4, finalPage);
             if (cmds.length !== 0) {
                 finalPage++;
             } else {
@@ -26,7 +27,7 @@ export const buttons: Buttons = {
         }
         finalPage -= 1;
 
-        const commands = commandPaginate([...client.commands.values()], 4, finalPage);
+        const commands = arrayPage([...client.commands.values()], 4, finalPage);
         const embed = new MessageEmbed()
             .setTitle(`${client.user?.tag}'s ${client.commands.size} Commands`)
             .setTimestamp()
