+++ src/events/ready.ts
@@ -1,36 +1,47 @@
 /* eslint-disable @typescript-eslint/member-ordering */
-import { CONFIG, STORAGE } from "../globals";
+import { CONFIG } from "../globals";
 import { Event } from "../interfaces/index";
 import { REST } from "@discordjs/rest";
 import { Routes } from "discord-api-types/v9";
+import { getGuild } from "../utils/getGuild";
+
 
 export const event: Event = {
     name: "ready",
     run: async (client) => {
-        console.log(`${client.user?.tag} is online!`);
-        console.log(STORAGE);
+        console.log(`${client.user?.tag} is online!\n`);
 
         if (!client.application?.owner) await client.application?.fetch();
 
-
+        if (client.application === null) {
+            throw new Error("Client Did not register in time, please try again");
+        }
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
         const commands = client.slashCommands.map(({ run, devOnly, guildOnly, dmOnly, ...data }) => data);
 
-
         try {
-            await client.application?.commands.set(commands);
+            if (CONFIG.devEnv.isDev) {
+                const guild = await getGuild(CONFIG.devEnv.devServer, client);
+
+                if (guild === null) {
+                    return void console.log("Could not find Dev ServerID");
+                }
+
+                await guild.commands.set(commands);
+                console.log("Set Commands for Dev Server\nCommands List:"
+                + `\n ${(await guild.commands.fetch()).map((c) => c.name)}\n`);
+            } else {
+                await client.application.commands.set(commands);
+                console.log("Set Commands for Production\nCommands List:"
+                + `\n ${(await client.application.commands.fetch()).map((c) => c.name)}\n`);
+            }
 
         } catch (error) {
             console.log(`There was an error registering a slash command \n${error}`);
         }
-        console.log(commands);
 
         const rest = new REST({ version: "9" }).setToken(CONFIG.token);
-        const clientID = client.application?.id;
-        if (clientID === undefined) {
-            throw new Error("There was an error getting the client ID");
-        }
-
+        const clientID = client.application.id;
 
         await (async (): Promise<void> => {
             try {
@@ -52,8 +63,7 @@ export const event: Event = {
 
                 }
 
-
-                console.log("Sucessfully reloaded application (/) commands.");
+                console.log("Sucessfully reloaded application (/) commands.\n");
             } catch (error) {
                 console.error(error);
             }
