+++ src/utils/getMember.ts
@@ -1,0 +1,18 @@
+import { Guild, GuildMember } from "discord.js";
+
+export async function getMember(uid: string, guild: Guild): Promise<GuildMember | null> {
+    let uidParsed = uid;
+    // Check if a member was tagged or not. If the member was tagged remove the
+    // Tag from uid.
+    if (uid.startsWith("<@") && uid.endsWith(">")) {
+        const re = new RegExp("[<@!>]", "g");
+        uidParsed = uid.replace(re, "");
+    }
+    // Try recovering the role and report if it was successful or not.
+    try {
+        return await guild.members.fetch(uidParsed);
+    } catch (e) {
+        console.log(`Member not found because ${e}`);
+        return null;
+    }
+}
