+++ src/utils/getRole.ts
@@ -1,0 +1,19 @@
+import { Guild, Role } from "discord.js";
+
+// Added getRole function in here incase you don't like getMember (;
+export function getRole(rid: string, guild: Guild): Role | undefined {
+    let ridParsed = rid;
+    // Check if a role was tagged or not. If the role was tagged remove the
+    // Tag from rid.
+    if (rid.startsWith("<@&") && rid.endsWith(">")) {
+        const re = new RegExp("[<@&>]", "g");
+        ridParsed = rid.replace(re, "");
+    }
+    // Try recovering the role and report if it was successful or not.
+    try {
+        return guild.roles.cache.get(ridParsed);
+    } catch (e) {
+        console.log(`Role not found because ${e}`);
+        return undefined;
+    }
+}
